+++ module-apps/application-calendar/models/DayEventsModel.hpp
@@ -4,20 +4,21 @@
 #include "Application.hpp"
 #include <ListItemProvider.hpp>
 
-class CalendarEventsModel : public gui::ListItemProvider
+class DayEventsModel : public gui::ListItemProvider
 {
     const int tempItemCount       = 5;
     app::Application *application = nullptr;
 
   public:
-    CalendarEventsModel(app::Application *app);
+    DayEventsModel(app::Application *app);
+    virtual ~DayEventsModel() override = default;
 
     void requestRecords(const uint32_t offset, const uint32_t limit) override;
 
     // virtual methods for ListViewProvider
-    unsigned int getMinimalItemHeight() const override;
+    [[nodiscard]] unsigned int getMinimalItemHeight() const override;
     gui::ListItem *getItem(gui::Order order) override;
-    int getItemCount() const override
+    [[nodiscard]] int getItemCount() const override
     {
         return tempItemCount;
     };